update: Double to BigDecimal

+++ src/main/java/com/rivernine/cryptoGeneratorBinance/schedule/analysis/AnalysisJob.java
@@ -1,9 +1,8 @@
 package com.rivernine.cryptoGeneratorBinance.schedule.analysis;
 
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Map;
 
-import com.rivernine.cryptoGeneratorBinance.client.model.trade.Order;
 import com.rivernine.cryptoGeneratorBinance.schedule.analysis.impl.AnalysisImpl;
 import com.rivernine.cryptoGeneratorBinance.schedule.market.dto.Candle;
 import com.rivernine.cryptoGeneratorBinance.schedule.market.dto.Symbol;
@@ -22,19 +21,19 @@ public class AnalysisJob {
     return analysisImpl.analysisCandles(candles, count);
   }
 
-  public String calAskPrice(Integer level, Symbol symbol, String coinQuantity, Double usedBalance) {
+  public String calAskPrice(Integer level, Symbol symbol, String coinQuantity, BigDecimal usedBalance) {
     return analysisImpl.calAskPrice(level, symbol, coinQuantity, usedBalance);
   }
 
-  public Double calLossCutPrice(String coinQuantity, Double usedBalance) {
+  public BigDecimal calLossCutPrice(String coinQuantity, BigDecimal usedBalance) {
     return analysisImpl.calLossCutPrice(coinQuantity, usedBalance);
   }
 
-  public Boolean judgeScaleTrade(Double curPrice, Double lastBidPrice, Integer level) {
+  public Boolean judgeScaleTrade(BigDecimal curPrice, BigDecimal lastBidPrice, Integer level) {
     return analysisImpl.judgeScaleTrade(curPrice, lastBidPrice, level);
   }
 
-  public String convertStepSize(Symbol symbol, Double quantity) {
+  public String convertStepSize(Symbol symbol, BigDecimal quantity) {
     return analysisImpl.convertStepSize(symbol, quantity);
   }
 }
