+++ src/main/java/com/rivernine/cryptoGeneratorBinance/schedule/market/dto/Candle.java
@@ -1,5 +1,7 @@
 package com.rivernine.cryptoGeneratorBinance.schedule.market.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,14 +13,14 @@ import lombok.ToString;
 public class Candle {
   private String symbol;
   private String time;
-  private Double open;      // 시가
-  private Double high;      // 고가
-  private Double low;       // 저가
-  private Double close;     // 종가
+  private BigDecimal open;      // 시가
+  private BigDecimal high;      // 고가
+  private BigDecimal low;       // 저가
+  private BigDecimal close;     // 종가
   private int flag;             // -1: 음봉, 0: 보합, 1: 양봉
 
   @Builder
-  public Candle(String symbol, String time, Double open, Double high, Double low, Double close, int flag){
+  public Candle(String symbol, String time, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close, int flag){
     this.symbol = symbol;
     this.time = time;
     this.open = open;
